vis mod with the no_pub macro param & better arg handling
Arguments handling of the macro have been improved. They can be supplied in any order now.

The visibility parameter no_pub has been implemented. Which switch off the all fields pub and let the user decide.
Moreover, the visibility of the struct is now correctly reflected in the generated code.